Started with code cleanup
